resolve circular dependency involving main by moving db services to a service file.  this also reduces the imports and parms for routers by leaning on a common sessiondep
