+++ src/main.py
@@ -1,7 +1,8 @@
-from fastapi import FastAPI, Depends, HTTPException
+from fastapi import FastAPI
 from fastapi_versioning import VersionedFastAPI, version
-from sqlmodel import Session, SQLModel, create_engine
-from typing import Annotated
+
+# service
+from src.services.db_service import create_db_and_tables
 
 
 # routers
@@ -18,26 +19,6 @@ from src.routers.persongoal_router import router as persongoal_router
 from src.routers.personskill_router import router as personskill_router
 from src.routers.personskillhistory_router import router as personskillhistory_router
 
-
-sqlite_file_name = "./data/database.db"
-sqlite_url = f"sqlite:///{sqlite_file_name}"
-
-connect_args = {"check_same_thread": False}
-engine = create_engine(sqlite_url, connect_args=connect_args)
-
-
-def create_db_and_tables():
-    SQLModel.metadata.create_all(engine)
-
-
-def get_session():
-    with Session(engine) as session:
-        yield session
-
-
-SessionDep = Annotated[Session, Depends(get_session)]
-
-
 base_app = FastAPI(title="Developer Developer")
 base_app.include_router(person_router)
 base_app.include_router(persongoal_router)
@@ -51,18 +32,6 @@ base_app.include_router(teamgoal_router)
 base_app.include_router(teamperson_router)
 base_app.include_router(tool_router)
 base_app.include_router(user_router)
-
-
-def connect_to_db():  # stub, inspiration
-    pass  # place holder
-
-
-def get_db():  # stub, inspiration
-    db = connect_to_db()
-    try:
-        yield db
-    finally:
-        db.close()
 
 
 @base_app.get("/")
