+++ src/routers/person_router.py
@@ -1,8 +1,7 @@
 from src.models.person import Person
 from src.services.person_service import create_person_service, read_person_service
-from src.main import SessionDep, get_session
-from generic_router_factory import create_crud_router
-
+from src.services.db_service import SessionDep
+from src.routers.generic_router_factory import create_crud_router
 
 
 router = create_crud_router(
@@ -11,5 +10,4 @@ router = create_crud_router(
     read_service=read_person_service,
     resource_name="persons",
     session_dep=SessionDep,
-    get_session_func=get_session,
 )
