+++ src/routers/team_router.py
@@ -1,7 +1,7 @@
 from src.models.team import Team
 from src.services.team_service import create_team_service, read_team_service
-from src.main import SessionDep, get_session
-from generic_router_factory import create_crud_router
+from src.services.db_service import SessionDep
+from src.routers.generic_router_factory import create_crud_router
 
 
 router = create_crud_router(
@@ -10,5 +10,4 @@ router = create_crud_router(
     read_service=read_team_service,
     resource_name="teams",
     session_dep=SessionDep,
-    get_session_func=get_session,
 )
