+++ src/routers/teamperson_router.py
@@ -1,7 +1,7 @@
 from src.models.teamperson import TeamPerson
 from src.services.teamperson_service import create_teamperson_service, read_teamperson_service
-from src.main import SessionDep, get_session
-from generic_router_factory import create_crud_router
+from src.services.db_service import SessionDep
+from src.routers.generic_router_factory import create_crud_router
 
 
 router = create_crud_router(
@@ -10,5 +10,4 @@ router = create_crud_router(
     read_service=read_teamperson_service,
     resource_name="teampersons",
     session_dep=SessionDep,
-    get_session_func=get_session,
 )
